remove CallProcessor (WIP)
