+++ wakepy/core/__init__.py
@@ -6,15 +6,13 @@ See the public Python API at: https://wakepy.readthedocs.io/
 
 from .activation import ActivationResult as ActivationResult
 from .activation import MethodActivationResult as MethodActivationResult
-from .calls import Call as Call
-from .calls import CallProcessor as CallProcessor
-from .calls import DbusMethodCall as DbusMethodCall
 from .constants import BusType as BusType
 from .constants import ModeName as ModeName
 from .constants import PlatformName as PlatformName
 from .dbus import DbusAdapter as DbusAdapter
 from .dbus import DbusAddress as DbusAddress
 from .dbus import DbusMethod as DbusMethod
+from .dbus import DbusMethodCall as DbusMethodCall
 from .method import Method as Method
 from .mode import Mode as Mode
 from .platform import CURRENT_PLATFORM as CURRENT_PLATFORM
