+++ wakepy/io/dbus/jeepney.py
@@ -2,8 +2,7 @@ from jeepney import DBusAddress, new_method_call
 from jeepney.io.blocking import open_dbus_connection
 from jeepney.wrappers import unwrap_msg
 
-from wakepy.core.calls import DbusMethodCall
-from wakepy.core.dbus import DbusAdapter
+from wakepy.core import DbusAdapter, DbusMethodCall
 
 
 class JeepneyDbusAdapter(DbusAdapter):
